Extract aliases config
